+++ apps/assets/src/components/asset/table/ContentTable.tsx
@@ -158,7 +158,7 @@ const ContentTable = ({
             <RowContent
               symbol={v.name}
               imgSrc={`/assets/tokenIdentifier/${v.icon.toLocaleLowerCase()}.png`}
-              valueInTokens={formatNumber(valueInTokens)}
+              valueInTokens={formatNumber(valueInTokens, 6)}
               valueInDollars={valueInDollars.toFixed(2)}
             />
           }