+++ CHANGELOG.md
@@ -35,10 +35,19 @@ Ref: https://keepachangelog.com/en/1.0.0/
 # Changelog
 
 ## Unreleased
+
+- (bug fixes) #fse-142 | apps/assets 1.0.1 | Change amount of decimals displayed: 6 instead of 2
+- (bug fixes) #fse-142 | apps/mission 1.0.1 | Display tooltip on topBarMissionControl with 6 decimals amount if amount is bigger than 0
+- (bug fixes) #fse-142 | apps/mission 1.0.1 | Change amount of decimals displayed: 6 instead of 2
+- (bug fixes) #fse-142 | apps/staking 1.0.1 | Change amount of decimals displayed: 6 instead of 2
+- (bug fixes) #fse-142 | apps/staking 1.0.1 | Display tooltip on topBarStaking with 6 decimals amount if amount is bigger than 0
+- (bug fixes) #fse-142 | packages/helpers 1.0.1 | Create function for display TopBar Tooltip
+- (chore) #fse-142 | evmos-wallet 1.0.2 | Export EVMOS_DECIMALS
 - (bug) #fse-400 | root config | Updating the dev workflow to avoid errors when running `yarn dev` 
 - (bug fixes) #fse-481 | evmos-wallet 1.0.1 | Show connected snackbar only if the user clicks on Connect Wallet
 - (ci) #fse-413 | (apps)/_ 1.0.x | Adding CI/CD pipeline for apps
 ...
+
 
 
 ## 1.0.0 - 2023-04-28